wip: new seeder config option
