+++ bin/cli.js
@@ -94,7 +94,11 @@ function invoke(env) {
       'environment, default: process.env.NODE_ENV || development'
     )
     .option('--esm', 'Enable ESM interop.')
-    .option('--specific [path]', 'Specify one seed file to execute.');
+    .option('--specific [path]', 'Specify one seed file to execute.')
+    .option(
+      '--timestamp-filename-prefix',
+      'Enable a timestamp prefix on name of generated seed files.'
+    );
 
   commander
     .command('init')
@@ -301,6 +305,11 @@ function invoke(env) {
       `--stub [<relative/path/from/knexfile>|<name>]`,
       'Specify the seed stub to use. If using <name> the file must be located in config.seeds.directory'
     )
+    .option(
+      '--timestamp-filename-prefix',
+      'Enable a timestamp prefix on name of generated seed files.',
+      false
+    )
     .action(async (name) => {
       const opts = commander.opts();
       opts.client = opts.client || 'sqlite3'; // We don't really care about client when creating seeds
@@ -310,6 +319,10 @@ function invoke(env) {
       const stub = getStubPath('seeds', env, opts);
       if (stub) {
         configOverrides.stub = stub;
+      }
+
+      if (opts.timestampFilenamePrefix) {
+        configOverrides.timestampFilenamePrefix = opts.timestampFilenamePrefix;
       }
 
       instance.seed
