+++ lib/migrate/MigrationGenerator.js
@@ -2,6 +2,7 @@ const path = require('path');
 const { writeJsFileUsingTemplate } = require('../util/template');
 const { getMergedConfig } = require('./configuration-merger');
 const { ensureDirectoryExists } = require('../util/fs');
+const { yyyymmddhhmmss } = require('../util/timestamp');
 
 class MigrationGenerator {
   constructor(migrationConfig) {
@@ -77,26 +78,6 @@ class MigrationGenerator {
       return path.resolve(process.cwd(), directory);
     });
   }
-}
-
-// Ensure that we have 2 places for each of the date segments.
-function padDate(segment) {
-  segment = segment.toString();
-  return segment[1] ? segment : `0${segment}`;
-}
-
-// Get a date object in the correct format, without requiring a full out library
-// like "moment.js".
-function yyyymmddhhmmss() {
-  const d = new Date();
-  return (
-    d.getFullYear().toString() +
-    padDate(d.getMonth() + 1) +
-    padDate(d.getDate()) +
-    padDate(d.getHours()) +
-    padDate(d.getMinutes()) +
-    padDate(d.getSeconds())
-  );
 }
 
 module.exports = MigrationGenerator;
