+++ lib/seed/Seeder.js
@@ -6,6 +6,7 @@ const extend = require('lodash/extend');
 const includes = require('lodash/includes');
 const { readdir, ensureDirectoryExists } = require('../util/fs');
 const { writeJsFileUsingTemplate } = require('../util/template');
+const { yyyymmddhhmmss } = require('../util/timestamp');
 
 // The new seeds we're performing, typically called from the `knex.seed`
 // interface on the main `knex` object. Passes the `knex` instance performing
@@ -83,6 +84,11 @@ class Seeder {
 
   _getNewStubFileName(name) {
     if (name[0] === '-') name = name.slice(1);
+
+    if (this.config.timestampFilenamePrefix === true) {
+      name = yyyymmddhhmmss() + '_' + name;
+    }
+
     return name + '.' + this.config.extension;
   }
 
@@ -148,6 +154,7 @@ class Seeder {
           '.ls',
           '.ts',
         ],
+        timestampFilenamePrefix: false,
       },
       this.config || {},
       config
